Initialize igraph in edges_to_graph.
This is the only place where we convert from non-igraph to igraph data
structures, so it's possible that a user calls this without knowing that they
are actually using igraph at all. The other functions accept (rai)igraph
objects so the library should already be initialized at that point.

We remove the igraph_bool_t from the function signature so that it's a
guaranteed that this function can be called without any knowledge of igraph.
Otherwise, it's possible that the DIRECTED/UNDIRECTED values are not actually
defined before the library is initialized.